Refactor messaging architecture: reorganize files, update imports, remove legacy code
