+++ src/application/TaskManager.cpp
@@ -7,8 +7,8 @@
 #include "../hardware/DeviceManager.h"
 #include "../hardware/SDManager.h"
 #include "../messaging/MessageAPI.h"
-#include "../messaging/InterruptMessagingEngine.h"
-#include "../messaging/MessageData.h"
+#include "../messaging/transport/SerialEngine.h"
+#include "../messaging/protocol/MessageData.h"
 #include "AppController.h"
 #include "AudioUI.h"
 #include "LogoSupplier.h"
