+++ src/messaging/protocol/MessageConfig.cpp
