+++ src/messaging/system/MessageConverter.cpp
@@ -1,4 +1,4 @@
-#include "MessageData.h"
+#include "../protocol/MessageData.h"
 #include <MessageProtocol.h>  // Direct import instead of relative path
 #include <esp_log.h>
