+++ src/messaging/transport/BinaryProtocol.cpp
