+++ src/messaging/transport/SerialEngine.cpp
@@ -1,9 +1,9 @@
-#include "InterruptMessagingEngine.h"
-#include "MessageAPI.h"
-#include "MessageConfig.h"
-#include "MessageData.h"
-#include "../include/MessagingConfig.h"
-#include "../include/BinaryProtocol.h"
+#include "SerialEngine.h"
+#include "../MessageAPI.h"
+#include "../protocol/MessageConfig.h"
+#include "../protocol/MessageData.h"
+#include "../../include/MessagingConfig.h"
+#include "../../include/BinaryProtocol.h"
 #include "../include/CoreLoggingFilter.h"
 #include "../include/DebugUtils.h"
 #include "../ui/screens/ui_screenDebug.h"
